+++ src/validate/checks/index.js
@@ -1,0 +1,7 @@
+module.exports = [
+  require('./overlay-files'),
+  require('./overlay-locales'),
+  require('./overlay-shape'),
+  require('./overlay-targets'),
+  require('./required-locales'),
+];